I added better styling to the view using the style guide provided by atom.
I also changed the behavior. I have a method which toggles the options on and off. The regular run insures that this is toggled off. When the option command is run it runs with this option toggled on. This should make sure that there is one script view per code file but allow easy changing between seeing it and not and allow for creation of a close button.

I am working on passing the values of the input into the option framework that I setup. The problem that I have right now is that when I type something into a text input, I cannot backspace the input. I don't know why this is happening but it seems related to a text input element. Once that is fixed, it should be easy to pass the arguments collected for configuration, add a save/close button, and be ready for some testing