Add Java grammar.
File-based Java scripts only for now. Selection-based scripts would
require naming the temp file to match the main class, and also finding
the main class in the selection. This could be done by passing in a file
name and searching for the first instance of /class (.*)\{/, but I don't
feel comfortable doing this without significant testing and suggestions
for the current contributors.