Allow using nested quotes and parentheses in Run Options
These changes allow using the expression shown in the added test, which allows
running racket scripts. This usage differs from the default racket behavior,
because this also outputs the return value of the script.

The change to `matches` creation uses a regex backreference and the lazy
(non-greedy) quantifier to only match outer quotes. So whereas the current
version would've matched both `"{FILE_ACTIVE}"` and `'(load "{FILE_ACTIVE}")'`,
then the new version only matches the outer quotes. The following code then
broke with the previous version, trying to match and replace both of these
overlapping instances.

The `args.replace` change was necessary, because otherwise the argument was
used as regex, and any parentheses and other regex operators were not being
escaped. Using simple substring replacing instead avoids the need to escape
special characters.

The last change avoids removing all quotes from the arguments and instead only
replaces matching quotes that are at the beginning and end of the argument.

I'm not sure if the "strips quotes from argument values" cases were actually
useful in any situation or the behavior was simply incidental. Currently
removed support for that behavior. If this behavior turns out to be necessary,
then separate handling should be added for it.