Fix and improve quote support in Run Options (#1424)
* Fix using quotes in Run Options

Currently, whenever Command Arguments or Program Arguments contain quotes in
the Run Options dialog, then the script blows up with a "Uncaught TypeError:
Cannot read property 'concat' of undefined".

Commit 668057c ("Make linter happy and fix convert errors", 2016-12-05) changed
`(matches != null) ? matches : []` to `!matches ? matches : []`, which reverses
the original meaning. This only causes errors when the string contains quotes,
because there's an early return in this function.

Fixes #1339.

* Test argument splitting

PR #742, that changed the logic from the simple "split on space" logic to the
current version, mentions these cases as the ones that were tested with this
implementation. Add tests to guarantee behavior remains the same for these
cases.

* Allow using nested quotes and parentheses in Run Options

These changes allow using the expression shown in the added test, which allows
running racket scripts. This usage differs from the default racket behavior,
because this also outputs the return value of the script.

The change to `matches` creation uses a regex backreference and the lazy
(non-greedy) quantifier to only match outer quotes. So whereas the current
version would've matched both `"{FILE_ACTIVE}"` and `'(load "{FILE_ACTIVE}")'`,
then the new version only matches the outer quotes. The following code then
broke with the previous version, trying to match and replace both of these
overlapping instances.

The `args.replace` change was necessary, because otherwise the argument was
used as regex, and any parentheses and other regex operators were not being
escaped. Using simple substring replacing instead avoids the need to escape
special characters.

The last change avoids removing all quotes from the arguments and instead only
replaces matching quotes that are at the beginning and end of the argument.

I'm not sure if the "strips quotes from argument values" cases were actually
useful in any situation or the behavior was simply incidental. Currently
removed support for that behavior. If this behavior turns out to be necessary,
then separate handling should be added for it.

* Simplify and improve argument splitting logic

Instead of multiple back and forth replacements, this uses a single regex to
find the arguments and simply adds them together while looping over the
provided text.

As far as I can tell, this now behaves exactly like most shells do. Except for
escaped quotes. This does not support escaping quotes.

This re-adds support for the "strips quotes from argument values" tests that
was removed in the previous commit c4c48cb ("Allow using nested quotes and
parentheses in Run Options", 2017-09-17).